+++ Core/OfficeDevPnP.Core/ALM/AppManager.cs
@@ -8,13 +8,14 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 
 namespace OfficeDevPnP.Core.ALM
 {
-#if !ONPREMISES
+#if !SP2013 && !SP2016
     /// <summary>
     /// Allows Application Lifecycle Management for Apps
     /// </summary>
@@ -509,6 +510,60 @@ namespace OfficeDevPnP.Core.ALM
         }
 
         /// <summary>
+        /// Synchronize an app from the tenant app catalog with the teams app catalog
+        /// </summary>
+        /// <param name="id">The unique id of the app. Notice that this is not the product id as listen in the app catalog</param>
+        /// <returns></returns>
+        public async Task<bool> SyncToTeamsAsync(Guid id)
+        {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentException(nameof(id));
+            }
+
+            await new SynchronizationContextRemover();
+
+            return await SyncToTeamsImplementation(id);
+        }
+
+        /// <summary>
+        /// Synchronize an app from the tenant app catalog with the teams app catalog
+        /// </summary>
+        /// <param name="appMetadata">The app metadata object of the app to remove.</param>
+        /// <returns></returns>
+        public async Task<bool> SyncToTeamsAsync(AppMetadata appMetadata)
+        {
+            if (appMetadata == null || appMetadata.Id == null)
+            {
+                throw new ArgumentException(nameof(appMetadata));
+            }
+
+            await new SynchronizationContextRemover();
+
+            return await SyncToTeamsImplementation(appMetadata.Id);
+        }
+
+        /// <summary>
+        /// Synchronize an app from the tenant app catalog with the teams app catalog
+        /// </summary>
+        /// <param name="id">The unique id of the app. Notice that this is not the product id as listen in the app catalog</param>
+        /// <returns></returns>
+        public bool SyncToTeams(Guid id)
+        {
+            return Task.Run(() => SyncToTeamsAsync(id)).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// Synchronize an app from the tenant app catalog with the teams app catalog
+        /// </summary>
+        /// <param name="appMetadata">The app metadata object of the app to remove.</param>
+        /// <returns></returns>
+        public bool SyncToTeams(AppMetadata appMetadata)
+        {
+            return Task.Run(() => SyncToTeamsAsync(appMetadata)).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
         /// Returns all available apps.
         /// </summary>
         /// <param name="scope">Specifies the app catalog to work with. Defaults to Tenant</param>
@@ -617,6 +672,13 @@ namespace OfficeDevPnP.Core.ALM
                     {
                         request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
                     }
+                    else
+                    {
+                        if (_context.Credentials is NetworkCredential networkCredential)
+                        {
+                            handler.Credentials = networkCredential;
+                        }
+                    }
                     request.Headers.Add("X-RequestDigest", await _context.GetRequestDigest());
 
                     // Perform actual post operation
@@ -689,16 +751,24 @@ namespace OfficeDevPnP.Core.ALM
                     handler.SetAuthenticationCookies(context);
                 }
 
+
                 using (var httpClient = new PnPHttpProvider(handler))
                 {
                     var method = action.ToString();
-                    string requestUrl = $"{context.Web.Url}/_api/web/{(scope == AppCatalogScope.Tenant ? "tenant" : "sitecollection")}appcatalog/AvailableApps/GetByID('{id}')/{method}";
+                    var requestUrl = $"{context.Web.Url}/_api/web/{(scope == AppCatalogScope.Tenant ? "tenant" : "sitecollection")}appcatalog/AvailableApps/GetByID('{id}')/{method}";
 
                     HttpRequestMessage request = new HttpRequestMessage(HttpMethod.Post, requestUrl);
                     request.Headers.Add("accept", "application/json;odata=nometadata");
                     if (!string.IsNullOrEmpty(accessToken))
                     {
                         request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
+                    }
+                    else
+                    {
+                        if (context.Credentials is NetworkCredential networkCredential)
+                        {
+                            handler.Credentials = networkCredential;
+                        }
                     }
                     request.Headers.Add("X-RequestDigest", await context.GetRequestDigest());
 
@@ -733,6 +803,84 @@ namespace OfficeDevPnP.Core.ALM
                     {
                         // Something went wrong...
                         throw new Exception(await response.Content.ReadAsStringAsync());
+                    }
+                }
+            }
+            return await Task.Run(() => returnValue);
+        }
+
+        private async Task<bool> SyncToTeamsImplementation(Guid appId)
+        {
+            var context = _context;
+
+            // switch context to appcatalog
+            var appcatalogUri = _context.Web.GetAppCatalog();
+            context = context.Clone(appcatalogUri);
+
+            var returnValue = false;
+            var accessToken = context.GetAccessToken();
+
+            using (var handler = new HttpClientHandler())
+            {
+                context.Web.EnsureProperty(w => w.Url);
+
+                // find the app by id
+
+                var list = context.Web.GetListByUrl("appcatalog");
+                var query = new CamlQuery();
+                query.ViewXml = $"<View><Query><Where><Contains><FieldRef Name='UniqueId'/><Value Type='Text'>{appId}</Value></Contains></Where></Query></View>";
+                var items = list.GetItems(query);
+                context.Load(items);
+                context.ExecuteQueryRetry();
+                
+                // we're not in app-only or user + app context, so let's fall back to cookie based auth
+                if (String.IsNullOrEmpty(accessToken))
+                {
+                    handler.SetAuthenticationCookies(context);
+                }
+                if (items.Count > 0)
+                {
+                    using (var httpClient = new PnPHttpProvider(handler))
+                    {
+                        var requestUrl = $"{context.Web.Url}/_api/web/tenantappcatalog/SyncSolutionToTeams(id={items[0].Id})";
+
+                        HttpRequestMessage request = new HttpRequestMessage(HttpMethod.Post, requestUrl);
+                        request.Headers.Add("accept", "application/json;odata=nometadata");
+                        if (!string.IsNullOrEmpty(accessToken))
+                        {
+                            request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
+                        }
+                        else
+                        {
+                            if (context.Credentials is NetworkCredential networkCredential)
+                            {
+                                handler.Credentials = networkCredential;
+                            }
+                        }
+                        request.Headers.Add("X-RequestDigest", await context.GetRequestDigest());
+
+                        // Perform actual post operation
+                        HttpResponseMessage response = await httpClient.SendAsync(request, new System.Threading.CancellationToken());
+
+                        if (response.IsSuccessStatusCode)
+                        {
+                            // If value empty, URL is taken
+                            var responseString = await response.Content.ReadAsStringAsync();
+                            if (responseString != null)
+                            {
+                                try
+                                {
+                                    var responseJson = JObject.Parse(responseString);
+                                    returnValue = true;
+                                }
+                                catch { }
+                            }
+                        }
+                        else
+                        {
+                            // Something went wrong...
+                            throw new Exception(await response.Content.ReadAsStringAsync());
+                        }
                     }
                 }
             }
@@ -774,6 +922,13 @@ namespace OfficeDevPnP.Core.ALM
                     if (!string.IsNullOrEmpty(accessToken))
                     {
                         request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
+                    }
+                    else
+                    {
+                        if (context.Credentials is NetworkCredential networkCredential)
+                        {
+                            handler.Credentials = networkCredential;
+                        }
                     }
                     request.Headers.Add("X-RequestDigest", requestDigest);
                     request.Headers.Add("binaryStringRequestBody", "true");