+++ Core/OfficeDevPnP.Core/Pages/AvailableClientSideComponents.cs
@@ -1,6 +1,6 @@
 ﻿namespace OfficeDevPnP.Core.Pages
 {
-#if !ONPREMISES
+#if !SP2013 && !SP2016
     /// <summary>
     /// Class holding a collection of client side webparts (retrieved via the _api/web/GetClientSideWebParts REST call)
     /// </summary>