+++ Core/OfficeDevPnP.Core/Pages/CanvasSectionTemplate.cs
@@ -1,6 +1,6 @@
 ﻿namespace OfficeDevPnP.Core.Pages
 {
-#if !ONPREMISES
+#if !SP2013 && !SP2016
     /// <summary>
     /// The type of canvas being used
     /// </summary>