+++ Core/OfficeDevPnP.Core/Pages/ClientSideText.cs
@@ -11,7 +11,7 @@ using System.Web.UI;
 
 namespace OfficeDevPnP.Core.Pages
 {
-#if !ONPREMISES
+#if !SP2013 && !SP2016
     /// <summary>
     /// Controls of type 4 ( = text control)
     /// </summary>