+++ Core/OfficeDevPnP.Core/Pages/ClientSideTextControlData.cs
@@ -2,7 +2,7 @@
 
 namespace OfficeDevPnP.Core.Pages
 {
-#if !ONPREMISES
+#if !SP2013 && !SP2016
 
     /// <summary>
     /// Control data for controls of type 4 (= text control)